pyproject.toml
[tool.poetry]
name = "matrixballsorter"
version = "0.1.0"
description = ""
authors = ["Declan Ketchum <dketchum@olin.edu>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
Pillow = "^8.4.0"
fastapi = "^0.70.0"
numpy = "^1.21.4"

[tool.poetry.dev-dependencies]
uvicorn = "^0.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
